+++ proto-lens-protoc/src/Data/ProtoLens/Compiler/Definitions.hs
@@ -72,6 +72,7 @@ import Proto.Google.Protobuf.Descriptor
     , FieldDescriptorProto'Type(..)
     , FileDescriptorProto
     , MethodDescriptorProto
+    , MethodOptions
     , ServiceDescriptorProto
     )
 import Data.ProtoLens.Compiler.Combinators
@@ -136,6 +137,7 @@ data MethodInfo = MethodInfo
     , methodOutput :: Text
     , methodClientStreaming :: Bool
     , methodServerStreaming :: Bool
+    , methodOptions :: MethodOptions
     }
 
 -- | Information about a single field of a proto message,
@@ -328,6 +330,7 @@ collectServices fd = fmap (toServiceInfo $ fd ^. #package) $ fd ^. #service
             , methodOutput = fromString . T.unpack $ md ^. #outputType
             , methodClientStreaming = md ^. #clientStreaming
             , methodServerStreaming = md ^. #serverStreaming
+            , methodOptions = md ^. #options
             }
 
 messageAndEnumDefs ::